requirements.txt
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
bpyutils==0.5.8
build==1.2.2.post1
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.8
colorclass==2.2.2
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==5.1.5
django-hosts==6.0
django-cors-headers==4.3.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.2.1
django-prometheus==2.3.1
django-comments-xtd==2.10.2
docopt==0.6.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jalali_core==1.0.0
jdatetime==5.1.0
msgpack==1.1.0
packaging==24.2
pillow==11.1.0
pip-upgrader==1.4.15
pipupgrade==1.12.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==25.0.0
pyproject_hooks==1.2.0
pytz==2024.2
redis==5.2.1
requests==2.32.3
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
terminaltables==3.1.10
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wheel==0.45.1
zope.interface==7.2

# Missing dependencies that are required by the application
celery==5.3.4
PyJWT==2.8.0
whitenoise==6.6.0
# Use pre-compiled wheels for Python 3.12 compatibility
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.3.0
